third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-svg-no-intrinsic-size-expected.txt

Test createImageBitmap from a SVG image without intrinsic size and with zero size.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS createImageBitmap from a SVG without intrinsic size rejected: InvalidStateError: The image element contains an SVG image without intrinsic dimensions, and no resize options or crop region are specified.
PASS createImageBitmap from a SVG with zero size rejected: InvalidStateError: The image element contains an SVG image without intrinsic dimensions, and no resize options or crop region are specified.
PASS createImageBitmap from a zero size SVG with cropRect succeed
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS createImageBitmap from a zero size SVG with resize and cropRect succeed
PASS successfullyParsed is true

TEST COMPLETE
